Start refactoring plenum protocol tests [WIP]
Added a setUp method to run the protocol parsing before each test. The
protocols are parsed into a dictionary, so that different protocols can
be tested.

Separated the different fields of the protocol header into different
tests, one field in each test.

While testing, I found a test that expects an exception parsing the
booklet number. This was due to a faulty regex in the parsing code. I
fixed the regex, and then updated the test to expect the correct parsed
number.